vpt: make http port (80) configurable
This patch enables to set the VPT module to connect to the VPT devices
listening on non-standard ports (other then 80). This is useful for testing,
eg. when having virtual VPT devices which cannot be bound to the port 80 on
a testing server (where 80 is already used or would require additional
permissions).

Signed-off-by: Jan Viktorin <viktorin@rehivetech.com>