build: fix building with relative paths
The cmake definitions were changed to accept the environment variable
RELATIVE_PATH (instead of RELATIVE_PATHS) which contains the relative prefix
we want to use for the application. Also, the compilation scripts are updated
so the first argument they accept is the relative path prefix. The paths in
application now read the contents of the RELATIVE_PATH macro.

Compilation with relative path ../special/etc:

$ ./compile_x86.sh ../special/etc
$ ./compile_armv7.sh ../special/etc

The used paths for AdaApp.ini and for modules are printed to stderr during
startup.

Signed-off-by: Jan Viktorin <viktorin@rehivetech.com>