vpt: fix EUID computation for VPT
The orignal function (from git:b1973f4) was wrong with extracting
the sub-MAC for EUID. The substr accepts offset + _length_ (and not
start + end). The goal is to extract the lowest 6 chars from the ID
string in format "XXXXXX######" ('#' is a hexa-decimal char we want
to extract).

After euid_t extraction (from git:a452471), the conversion is
broken even more by adding auto-detection of the number format -
stoull(..., 16) -> stoull(..., 0). Now, we force the hexa decimal
format again.

Fixes: a4524712c6b5 ("euid: extract data type euid_t")
Fixes: b1973f45eee4 ("vpt: Add new universal module for...")

Acked-by: Jan Viktorin <iviktorin@fit.vutbr.cz>