main: split creating and starting of the threads
Lets do the full initialization of the modules first. Start them after
the initialization. This behaviour is better, because modules will be
fully initialized before starting the threads.

Some memory leaks are fixed as a side effect of this change. These
memory leaks were introduced in the previous commit (4994af80).

Signed-off-by: Tomas Novotny <tomas@novotny.cz>
Signed-off-by: Adam Trhon <adam.trhon@gmail.com>