vpt: add new attribute paired into VPTDevice
The attribute paired represents whether a detected VPT device is paired or not.
A new detected VPT is automatically marked as unpaired (false). If a detected
VPT is already known, we take the paired status cached in adaapp (that can be
either paired or unpaired).

We have to use map_devices[id] = ... syntax to overwrite the existing
value (insert would not overwrite it).

Acked-by: Jan Viktorin <iviktorin@fit.vutbr.cz>