vpt: add attribute active
Attribute active prevents access to paired devices which are not detected
on LAN. As we do not know their IP addresses, we cannot read their contents
anyway. The unpaired devices sends their data (if they are available) to
notify their existence to the server.

If a paired device is detected on LAN, its active state is changed to
ACTIVE (true). This enables retreiving data for this device.

Acked-by: Jan Viktorin <iviktorin@fit.vutbr.cz>