vpt: update paired devices on listen command
The listen command initiates a scan of paired devices on the network.
The scan is delayed by 3 minutes (approximate expiration of the listen process
on the gateway).

After the listen is done (max. 2 minutes long), the user selects
devices to pair. There is currently no message coming to the adapter
telling what was paired. So, we have to poll server for those info
after the listen command finishes.

Acked-by: Jan Viktorin <iviktorin@fit.vutbr.cz>