vpt: delete not paired devices only in pairedDevices
If there is no configuration for a device (meaning it is unsupported)
an exception is caught. In such case, the unpaired devices are deleted.

A paired device (according to data from server) can be missing on the LAN.
In that case, we do not have enough information about it and cannot load
the configuration either. Preserve such device until its timeout exceeds
or it appears on the network.

Acked-by: Jan Viktorin <iviktorin@fit.vutbr.cz>