PressureSensor: Added wakeup_value sending check
Obtaining new refresh value on startup is not always successful
because not all parts of the gateway are initialized. Due to this
fact, the PressureSensor cannot obtain the refresh time set by its
user and uses a default value that is then forced to the server.
Such behaviour is incorrect.

PressureSensor module now don't send the default value if there is
failure on obtaining the referesh time. It would repeatedly sending
requests for the refresh time value instead until it succeeds.

Signed-off-by: Richard Wolfert <xwolfe00@stud.fit.vutbr.cz>
[J. Viktorin: modified commit message]
Acked-by: Jan Viktorin <iviktorin@fit.vutbr.cz>