vpt: delete inactive devices automatically
A device is automatically deleted if it is not paired or if it is
considered INACTIVE.

The attribute active is extended to determine when a certain unpaired
device should be removed from cache. The attribute active now represents
time until the device is considered ACTIVE.

Each iteration (on wake up), the active timeout is decremented by
wake_up_time for every unpaired device. When the active timeout
reaches active <= INACTIVE for an unpaired device it can be removed.

For paired devices, we do not touch the active attribute and leave it
at the initial value.

The initialization value for each device is 5 minutes. This value was
derived from the pair mode time that last for approx 2 minutes. We will
use the 3 extra minutes in the following patches when automatically
deleting devices.

States:

* INACTIVE (active <= 0)
    paired device:   inactive, not deleted, not accessed
    unpaired device: inactive, can be deleted

* ACTIVE (active > 0)
    paired device:   active, not deleted, active is not
                     decremented, sensors are read
    unpaired device: active, not deleted, active is being
                     decremented, sensors are read

Acked-by: Jan Viktorin <iviktorin@fit.vutbr.cz>