LedModule: recover latest user configuration from server
Extend LedModule with the ability to recover the last configuration of
the module fom server. On start, the LedModule requests the last known
configuration (set by user) from server.

Due to this, default value is forced only when there are no
latest values on server. All leds are turned off by default.

Signed-off-by: Richard Wolfert <xwolfe00@stud.fit.vutbr.cz>
Acked-by: Jan Viktorin <iviktorin@fit.vutbr.cz>