aggregator: fix message propagating
Openhab prevented delivering messages to PanInterface.

Openhab was too high in the else if condition chain so no messages were propagated
to the PanInterface. The reason is that Openhab cannot decide which messages belong
to it so it accepts all.

Aggregator is now propagating messages to both of this modules, if no modules before
handled them.

Fixes: f8859b923c81 ("protocol: add support for version 1.1, add messages getparameter/parameter")

Signed-off-by: Richard Wolfert <xwolfe00@stud.fit.vutbr.cz>
Acked-by: Jan Viktorin <iviktorin@fit.vutbr.cz>