vpt: fix search of device specification
Support of different web pages versions requires to load the correct specification.
The name of a specification file is a concatenation of the target VPT's name and version.

This patch fixes setting of VPT's actuators which didn't work
after the specification version support was introduced.
Loading the specification file by device name only is not sufficient here.

Fixes: d376a71 ("vpt: add support for diffrent web pages")

Reported-by: Jan Viktorin <iviktorin@fit.vutbr.cz>
Signed-off-by: Lukas Koszegy <l.koszegy@gmail.com>
Acked-by: Jan Viktorin <iviktorin@fit.vutbr.cz>