utils: new attributes for WebSocketServerConnection
Because of using Websocket communication, protocol needed to be updated
with new message types and attributes for managing responses/requests.
"IOTMessage" and "Command" structs were encapsulated in new structures
"ServerMessage" and "ServerCommand" respectively to be able to hold these attributes.
Due to this, Distributor, TCP and xmltool modules were updated.

Signed-off-by: Richard Wolfert <xwolfe00@stud.fit.vutbr.cz>
Reviewed-by: Jan Viktorin <iviktorin@fit.vutbr.cz>